test(activesupport): move trails-only ParameterFilter cases to the .trails sibling
`parameter-filter.test.ts` maps to
`vendor/rails/activesupport/test/parameter_filter_test.rb`, so it should hold
one describe per Rails test class. It carried a second
`describe("ParameterFilter")` of six trails-only cases with no Rails
counterpart; those move into the existing `parameter-filter.trails.test.ts`,
per CLAUDE.md's rule that TS-only extras live in the `.trails` sibling.

The duplicate `describe` blocks the story also cites in `key-generator.test.ts`
(plus its misplaced BacktraceCleaner copies) were already removed on main.
