+++ packages/activesupport/src/parameter-filter.test.ts
@@ -191,52 +191,3 @@ describe("ParameterFilterTest", () => {
     expect(precompiled.filter((filter) => procs.includes(filter as () => void))).toEqual(procs);
   });
 });
-
-describe("ParameterFilter", () => {
-  it("recurses into nested plain objects", () => {
-    const filter = new ParameterFilter(["secret"]);
-    const result = filter.filter({ outer: { secret: "hidden", public: "visible" } });
-    expect((result.outer as any).secret).toBe("[FILTERED]");
-    expect((result.outer as any).public).toBe("visible");
-  });
-
-  it("recurses into arrays", () => {
-    const filter = new ParameterFilter(["secret"]);
-    const result = filter.filter({ items: [{ secret: "a" }, { secret: "b", x: 1 }] });
-    expect((result.items as any[])[0].secret).toBe("[FILTERED]");
-    expect((result.items as any[])[1].secret).toBe("[FILTERED]");
-    expect((result.items as any[])[1].x).toBe(1);
-  });
-
-  it("preserves Date instances without corruption", () => {
-    const filter = new ParameterFilter(["secret"]);
-    const d = new Date("2024-01-15T10:00:00.000Z");
-    expect(filter.filterParam("created_at", d)).toBe(d);
-  });
-
-  it("preserves non-plain class instances without corruption", () => {
-    class Foo {
-      constructor(public val: number) {}
-    }
-    const filter = new ParameterFilter(["secret"]);
-    const obj = new Foo(42);
-    expect(filter.filterParam("foo", obj)).toBe(obj);
-  });
-
-  it("filters null-prototype objects", () => {
-    const filter = new ParameterFilter(["secret"]);
-    const params = Object.assign(Object.create(null), { secret: "hidden", name: "alice" });
-    const result = filter.filter(params as Record<string, unknown>);
-    expect(result.secret).toBe("[FILTERED]");
-    expect(result.name).toBe("alice");
-  });
-
-  it("filters a nested HashWithIndifferentAccess in place", () => {
-    const filter = new ParameterFilter(["secret"]);
-    const params = withIndifferentAccess({ outer: withIndifferentAccess({ secret: "hidden" }) });
-    const result = filter.filter(params as unknown as Record<string, unknown>);
-    expect((result as unknown as HashWithIndifferentAccess<any>).get("outer").get("secret")).toBe(
-      "[FILTERED]",
-    );
-  });
-});
